Replace logger for newer DIETLogger method
