+++ src/main/java/me/besser/TAPI.java
@@ -6,25 +6,22 @@ import org.bukkit.plugin.java.JavaPlugin;
 import net.milkbowl.vault.economy.Economy;
 import org.bukkit.plugin.RegisteredServiceProvider;
 
-import static java.util.logging.Level.*;
-import static me.besser.TAPILogger.log;
+import static me.besser.DIETLogger.*;
 
 public final class TAPI extends JavaPlugin {
     private static Economy econ = null;
 
     @Override
     public void onEnable() {
-        TAPILogger.initialize(this);
+        DIETLogger.initialize(this);
 
         boolean isEnabledInConfig = getConfig().getBoolean("tapi.enable", true);
         if (!isEnabledInConfig) {
-            log(WARNING, "TAPI is disabled in config.yml and will not be fully initialized");
+            log(WARNING, "TAPI is disabled in config.yml and will not start");
             getServer().getPluginManager().disablePlugin(this);
             return;
         }
 
-        // Also kind of relies on essentials, try and see if that's needed too.
-        // This might not be needed due to the dependency check in plugin.yml
         if (!setupEconomy()) {
             log(SEVERE, "Vault is not found! This may cause bugs when responding to API requests");
             return;
@@ -32,7 +29,7 @@ public final class TAPI extends JavaPlugin {
 
         saveDefaultConfig();
 
-        PlayerTracker playerTracker = new PlayerTracker();
+        PlayerTracker playerTracker = new PlayerTracker(this);
         getServer().getPluginManager().registerEvents(playerTracker, this);
 
         new PlayerDataServer(this, playerTracker);
@@ -44,6 +41,7 @@ public final class TAPI extends JavaPlugin {
         if (getServer().getPluginManager().getPlugin("Vault") == null) {
             return false;
         }
+
         RegisteredServiceProvider<Economy> rsp = getServer().getServicesManager().getRegistration(Economy.class);
         if (rsp == null) {
             return false;
